+++ CHANGES.md
@@ -1,5 +1,11 @@
 # **Pantheios.Ruby** Changes
 
+## 0.21.0 - 3rd June 2020
+
+* added `Pantheios::ApplicationLayer::StockSeverityLevels::STOCK_SEVERITY_LEVEL_ALIASES`, which is a map that provides lookup of all recognised severity levels and aliases to the canonical severity level symbol
+* added `:debug5` severity level
+* now dependent on **xqsr3** v0.36+
+
 ## 0.20.3.2 - 3rd June 2020
 
 * more-complete tests for stock severity levels